Projects/Persona/python/requirements.txt
supabase>=2.0.0
psutil>=5.9.0
click>=8.1.0
tabulate>=0.9.0
python-dotenv>=1.0.0
pydantic>=2.0.0
watchdog>=3.0.0
